refactor: deepen public asset delivery into one tested HTTP boundary
The precondition/range evaluator and its decision union fold into
delivery.ts as private implementation; the module's surface is the
handler, the path helper, and the two ports. The boundary suite now
proves final responses through the production handler: status, headers,
bodies, provider call counts, cache population and parity, sanitized
failures, date preconditions, If-Range downgrade, and stale-body
cancellation. The decision-level test file retires with its
status-only assertions migrated.

Closes #202
